Refactor lobby list observer: Rename methods and parameters for consistency

+++ src/main/java/it/polimi/ingsw/gc20/server/model/ship/Ship.java
@@ -116,7 +116,7 @@ public abstract class Ship {
             if (cannon.getRotation() == Direction.UP)
                 power += cannon.getPower();
             else {
-                power += cannon.getPower() / 2;
+                power += (float) cannon.getPower() / 2;
             }
         }
         return power;
